pyproject.toml
[tool.poetry]
name = "image_generator_bot"
version = "0.1.0"
description = ""
authors = ["SayaleeDamle <sayalee.damle@onepointltd.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.313"
openai = "^0.28.1"
chainlit = "^0.7.2"
python-dotenv = "^1.0.0"
faiss-cpu = "^1.7.4"
tiktoken = "^0.5.1"
pydantic = "^2.4.2"
pandas = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
